+++ sumologic/resource_sumologic_azure_event_hub_log_source_test.go
@@ -41,7 +41,6 @@ func TestAccSumologicAzureEventHubLogSource_create(t *testing.T) {
 					resource.TestCheckResourceAttr(azureEventHubLogResourceName, "content_type", "AzureEventHubLog"),
 					resource.TestCheckResourceAttr(azureEventHubLogResourceName, "path.0.type", "AzureEventHubPath"),
 				),
-				ExpectNonEmptyPlan: true,
 			},
 		},
 	})
@@ -75,7 +74,6 @@ func TestAccSumologicAzureEventHubLogSource_update(t *testing.T) {
 					resource.TestCheckResourceAttr(azureEventHubLogResourceName, "content_type", "AzureEventHubLog"),
 					resource.TestCheckResourceAttr(azureEventHubLogResourceName, "path.0.type", "AzureEventHubPath"),
 				),
-				ExpectNonEmptyPlan: true,
 			},
 			{
 				Config: testAccSumologicAzureEventHubLogSourceConfig(cName, cDescription, cCategory, sNameUpdated, sDescriptionUpdated, sCategoryUpdated, testSASKeyName, testSASKey, testNamespace, testEventHub, testConsumerGroup, testRegion),
@@ -89,7 +87,6 @@ func TestAccSumologicAzureEventHubLogSource_update(t *testing.T) {
 					resource.TestCheckResourceAttr(azureEventHubLogResourceName, "content_type", "AzureEventHubLog"),
 					resource.TestCheckResourceAttr(azureEventHubLogResourceName, "path.0.type", "AzureEventHubPath"),
 				),
-				ExpectNonEmptyPlan: true,
 			},
 		},
 	})