+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ENHANCEMENTS:
 * Add support for monitor frequency configurability
+BUG FIXES:
+* Polling source paths without `use_versioned_api` will not show an incorrect diff in the `terraform plan`
 
 ## 3.0.4 (February 4, 2025)
 BUG FIXES: